requirements.txt
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
contextlib2==21.6.0
dominate==2.6.0
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
frozenlist==1.2.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.2.1
MarkupSafe==2.0.1
mygrations==0.13.0
mysql-connector-python==8.0.26
protobuf==3.19.1
pyrsistent==0.18.0
regex==2021.11.10
requests==2.26.0
schema==0.7.4
tqdm==4.51.0
urllib3==1.26.7
visitor==0.1.3
websocket-client==1.2.1
Werkzeug==2.0.2